+++ crates/tanoshi/src/presentation/graphql/library.rs
@@ -121,7 +121,7 @@ impl LibraryRoot {
                 }
 
                 let mut connection = Connection::new(has_previous_page, has_next_page);
-                connection.append(
+                connection.edges.extend(
                     edges
                         .into_iter()
                         .map(|e| Edge::new(Cursor(e.uploaded.timestamp(), e.chapter_id), e.into())),
@@ -186,7 +186,7 @@ impl LibraryRoot {
                 }
 
                 let mut connection = Connection::new(has_previous_page, has_next_page);
-                connection.append(
+                connection.edges.extend(
                     edges
                         .into_iter()
                         .map(|e| Edge::new(Cursor(e.read_at.timestamp(), e.manga_id), e.into())),